+++ bindings/nodejs/examples/how_tos/accounts_and_addresses/create_address.ts
@@ -26,7 +26,7 @@ async function run() {
         // To create an address we need to unlock stronghold.
         await wallet.setStrongholdPassword(process.env.STRONGHOLD_PASSWORD);
 
-        const address = (await account.generateAddresses(1))[0];
+        const address = (await account.generateEd25519Addresses(1))[0];
 
         console.log(`Generated address:`, address.address);
     } catch (error) {