+++ bindings/nodejs/lib/wallet/Account.ts
@@ -377,7 +377,7 @@ export class Account {
     async generateAddress(
         options?: GenerateAddressOptions,
     ): Promise<AccountAddress> {
-        const addresses = await this.generateAddresses(1, options);
+        const addresses = await this.generateEd25519Addresses(1, options);
         return addresses[0];
     }
 
@@ -387,14 +387,14 @@ export class Account {
      * @param options Options for address generation.
      * @returns The addresses.
      */
-    async generateAddresses(
+    async generateEd25519Addresses(
         amount: number,
         options?: GenerateAddressOptions,
     ): Promise<AccountAddress[]> {
         const response = await this.methodHandler.callAccountMethod(
             this.meta.index,
             {
-                name: 'generateAddresses',
+                name: 'generateEd25519Addresses',
                 data: {
                     amount,
                     options,