+++ bindings/python/examples/client/02_generate_addresses.py
@@ -43,13 +43,13 @@ print('List of generated internal addresses:', *addresses, sep='\n')
 print()
 
 # Generate addresses with providing all inputs, that way it can also be done offline without a node.
-addresses = client.generate_addresses(secret_manager,
+addresses = client.generate_ed25519_addresses(secret_manager,
                                       coin_type=CoinType.SHIMMER,
                                       account_index=0,
                                       start=0,
                                       end=4,
                                       internal=False,
-                                      # Generating addresses with client.generateAddresses(secretManager, options={}), will by default get the bech32_hrp (Bech32
+                                      # Generating addresses with client.generateEd25519Addresses(secretManager, options={}), will by default get the bech32_hrp (Bech32
                                       # human readable part) from the node info, generating it "offline" requires setting it in the generateAddressesOptions
                                       bech32_hrp='rms')