+++ sdk/src/client/bindings/java/examples/src/GetAddressBalance.java
@@ -19,19 +19,21 @@ import java.util.Map;
 public class GetAddressBalance {
     public static void main(String[] args) throws ClientException, InitializeClientException {
         // Build the client.
-        Client client = new Client(new ClientConfig().withNodes(new String[]{"https://api.testnet.shimmer.network"}));
+        Client client = new Client(
+                new ClientConfig().withNodes(new String[] { "https://api.testnet.shimmer.network" }));
 
         // Generate the addresses from the given mnemonic.
-        MnemonicSecretManager secretManager = new MnemonicSecretManager("endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river");
-        String[] addresses = client.generateAddresses(secretManager, new GenerateAddressesOptions().withRange(new Range(0, 1)));
+        MnemonicSecretManager secretManager = new MnemonicSecretManager(
+                "endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river");
+        String[] addresses = client.generateEd25519Addresses(secretManager,
+                new GenerateAddressesOptions().withRange(new Range(0, 1)));
 
         // Get the interesting output ids.
         OutputId[] outputIds = client.getBasicOutputIds(new NodeIndexerApi.QueryParams()
                 .withParam("address", addresses[0])
                 .withParam("hasExpiration", false)
                 .withParam("hasTimelock", false)
-                .withParam("hasStorageDepositReturn", false)
-        ).getItems();
+                .withParam("hasStorageDepositReturn", false)).getItems();
 
         // Get the outputs.
         List<Map.Entry<Output, OutputMetadata>> outputs = client.getOutputs(outputIds);
@@ -43,12 +45,12 @@ public class GetAddressBalance {
             Output o = entry.getKey();
 
             if (o.toJson().has("nativeTokens")) {
-                for(JsonElement elem: o.toJson().get("nativeTokens").getAsJsonArray()) {
+                for (JsonElement elem : o.toJson().get("nativeTokens").getAsJsonArray()) {
                     NativeToken nativeToken = new NativeToken(elem.getAsJsonObject());
                     String tokenId = nativeToken.toJson().get("id").getAsString();
                     String amount = nativeToken.toJson().get("amount").getAsString().replace("0x", "");
 
-                    if(nativeTokens.containsKey(tokenId))
+                    if (nativeTokens.containsKey(tokenId))
                         nativeTokens.put(tokenId, nativeTokens.get(tokenId) + Integer.parseInt(amount, 16));
                     else
                         nativeTokens.put(tokenId, Integer.parseInt(amount, 16));